api/requirements.txt
langchain
python-dotenv
langchain-community
langchain-experimental
langchain-core
pydantic
pydantic-core
langchain_google_genai
langsmith
fastapi
langchain-pinecone
torch
sentence-transformers
pyyaml
langchain-huggingface
uvicorn
langchain_anthropic 
langchain_google_vertexai 